requirements.txt
google-cloud-bigquery==1.24.0
pandas==0.24.0